Move binary::main into actual main
